pyproject.toml
[project]
name = "chatbot-recruitment"
version = "0.1.0"
description = "Recruitment Assistant Chatbot using Agno and Streamlit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=2.0.0",
    "streamlit>=1.40.0",
    "python-dotenv>=1.0.0",
    "lancedb>=0.15.0",
    "gspread>=6.0.0",
    "oauth2client>=4.1.3",
    "googlesearch-python>=1.2.5",
    "pycountry>=24.6.1",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.44",
    "aiofiles>=25.1.0",
    "google-genai>=1.48.0",
    "rich>=14.2.0",
]

[tool.uv]
dev-dependencies = []

